added the frontend code
The frontend uses the fetch js package. You have to wait a few
seconds-minute for the Promise to resolve when the request is made to
the backend. until further changes, testing on the backend needs to be
done through the api docs: http://127.0.0.1:8000/docs, as the prompt
used on via the frontend is hard coded until further parsing of the
email body is done